Remove safe.engine.utilities
